requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.0
pyte>=0.8.0
